add test module for extractor

+++ data/extractor.py
@@ -2,10 +2,14 @@ import pathlib
 import shutil
 
 
+class ArchiveModifiedError(Exception):
+    pass
+
+
 class OpenArchive:
     tol = 1
 
-    def __init__(self, archive: pathlib.Path, tmpdir: str = "tmp"):
+    def __init__(self, archive: pathlib.Path, tmpdir: str = "_tmp"):
         self._archive = archive
         self._tmpdir = pathlib.Path(tmpdir)
 
@@ -28,35 +32,14 @@ class OpenArchive:
         num_at_exit = len(files)
         size_at_exit = sum([f.stat().st_size for f in files])
 
-        assert (
-            num_at_exit == self._num_at_enter
-        ), f"Number of files in temporary directory, {self._tmpdir} has changed since archive was extracted."
-        assert (
-            abs(size_at_exit - self._size_at_enter) < self.tol
-        ), f"Size of temporary directory, {self._tmpdir}, has changed since archive was extracted."
         # TODO log these as errors for investigation
+        if not num_at_exit == self._num_at_enter:
+            raise ArchiveModifiedError(
+                f"Number of files in temporary directory, {self._tmpdir} has changed since archive was extracted."
+            )
+        if not abs(size_at_exit - self._size_at_enter) < self.tol:
+            raise ArchiveModifiedError(
+                f"Size of temporary directory, {self._tmpdir}, has changed since archive was extracted."
+            )
 
         shutil.rmtree(self._tmpdir)
-
-
-# TODO put into test module
-def test_num_files_changes():
-    p = pathlib.Path("CORA-5.2-global-1950.tgz")
-    try:
-        with OpenArchive(p) as files:
-            f = next(files)
-            f.unlink()
-    except AssertionError as e:
-        print(e)
-
-
-def test_size_changes():
-    p = pathlib.Path("CORA-5.2-global-1950.tgz")
-    try:
-        with OpenArchive(p) as files:
-            f = next(files)
-            f.unlink()
-            g = next(files)
-            shutil.copy(g, f)
-    except AssertionError as e:
-        print(e)
